Remove dummy print statement, move expand_1d
